+++ src/contexts/EtherspotContext.tsx
@@ -1,12 +1,14 @@
 import { createContext } from 'react';
 import { DataUtils, PrimeSdk, WalletProviderLike } from '@etherspot/prime-sdk';
+import { ModularSdk, WalletProviderLike as WalletProviderLikeModular } from '@etherspot/modular-sdk';
 
 export interface EtherspotContextData {
   data: {
-    getSdk: (chainId?: number, forceNewInstance?: boolean) => Promise<PrimeSdk>;
+    getSdk: (modular?: boolean, chainId?: number, forceNewInstance?: boolean) => Promise<ModularSdk | PrimeSdk>;
     getDataService: () => DataUtils;
-    provider: WalletProviderLike | null | undefined;
+    provider: WalletProviderLike | WalletProviderLikeModular | null | undefined;
     chainId: number;
+    isModular: boolean;
   }
 }