+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.13.0] - 2024-06-17
+
+### Added Changes
+- `getSDK` include a param to choose to instantiate the Prime SDK instead of the Modular SDK
+- Added Etherspot Modular SDK `installModule` and `uninstallModule` to hook `useEtherspotModules`
+- `EtherspotContextProvider` updated with choice of instantiatiating the Prime SDK or Modular SDK
+- Added `isModular` to context `EtherspotContextProvider`
+- 
+
+### Breaking Changes
+- Etherspot Modular SDK implemented to Transaction Kit as the default SDK
+- The account template in the `EtherspotTransactionKit` component, can only be 'etherspot' if `moddular` is true.
+
 ## [0.12.1] - 2024-05-22
 
 ### Added Changes